@@ -10,7 +10,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class HeightCast implements CastsAttributes
 {
-    public function get(Model $model, string $key, mixed $value, array $attributes)
+    public function get(Model $model, string $key, mixed $value, array $attributes): Height
     {
         $feet = (int) floor($value / 12);
         $inches = $value % 12;
@@ -18,7 +18,7 @@ class HeightCast implements CastsAttributes
         return new Height($feet, $inches);
     }
 
-    public function set(Model $model, string $key, mixed $value, array $attributes)
+    public function set(Model $model, string $key, mixed $value, array $attributes): mixed
     {
         return $value;
     }
