@@ -27,14 +27,14 @@ class EventsList extends BaseComponent
      *
      * @var bool
      */
-    public $showFilters = false;
+    public bool $showFilters = false;
 
     /**
      * Shows list of accepted filters and direction to be displayed.
      *
      * @var array<string, string>
      */
-    public $filters = [
+    public array $filters = [
         'search' => '',
     ];
 
