@@ -18,7 +18,7 @@ use Tests\RequestFactories\StableRequestFactory;
 class StoreRequest extends FormRequest
 {
     /** @var class-string */
-    public static $factory = StableRequestFactory::class;
+    public static string $factory = StableRequestFactory::class;
 
     /**
      * Determine if the user is authorized to make this request.
