@@ -7,6 +7,7 @@ namespace App\Models\Concerns;
 use Ankurk91\Eloquent\HasBelongsToOne;
 use Ankurk91\Eloquent\Relations\BelongsToOne;
 use App\Models\TagTeam;
+use App\Models\TagTeamPartner;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 
 /**
@@ -24,7 +25,8 @@ trait CanJoinTagTeams
     public function tagTeams(): BelongsToMany
     {
         return $this->belongsToMany(TagTeam::class, 'tag_team_wrestler')
-            ->withPivot(['joined_at', 'left_at']);
+            ->withPivot(['joined_at', 'left_at'])
+            ->using(TagTeamPartner::class);
     }
 
     /**
