@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Models\Concerns;
 
+use App\Models\Managable;
 use App\Models\TagTeam;
 use App\Models\Wrestler;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
@@ -18,7 +19,8 @@ trait Manageables
     public function wrestlers(): MorphToMany
     {
         return $this->morphedByMany(Wrestler::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at']);
+            ->withPivot(['hired_at', 'left_at'])
+            ->using(Managable::class);
     }
 
     /**
@@ -28,8 +30,7 @@ trait Manageables
      */
     public function currentWrestlers(): MorphToMany
     {
-        return $this->morphedByMany(Wrestler::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at'])
+        return $this->wrestlers()
             ->wherePivotNull('left_at');
     }
 
@@ -40,8 +41,7 @@ trait Manageables
      */
     public function previousWrestlers(): MorphToMany
     {
-        return $this->morphedByMany(Wrestler::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at'])
+        return $this->wrestlers()
             ->wherePivotNotNull('left_at');
     }
 
@@ -53,7 +53,8 @@ trait Manageables
     public function tagTeams(): MorphToMany
     {
         return $this->morphedByMany(TagTeam::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at']);
+            ->withPivot(['hired_at', 'left_at'])
+            ->using(Managable::class);
     }
 
     /**
@@ -63,8 +64,7 @@ trait Manageables
      */
     public function currentTagTeams(): MorphToMany
     {
-        return $this->morphedByMany(TagTeam::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at'])
+        return $this->tagTeams()
             ->wherePivotNull('left_at');
     }
 
@@ -75,8 +75,7 @@ trait Manageables
      */
     public function previousTagTeams(): MorphToMany
     {
-        return $this->morphedByMany(TagTeam::class, 'manageable')
-            ->withPivot(['hired_at', 'left_at'])
+        return $this->tagTeams()
             ->wherePivotNotNull('left_at');
     }
 }
