@@ -12,7 +12,6 @@ use App\Models\Contracts\Employable;
 use App\Models\Contracts\Manageable;
 use App\Models\Contracts\Retirable;
 use App\Models\Contracts\Suspendable;
-use Fidum\EloquentMorphToOne\HasMorphToOne;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -20,6 +19,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 class TagTeam extends Model implements Bookable, CanBeAStableMember, Employable, Manageable, Retirable, Suspendable
 {
     use Concerns\CanJoinStables;
+    use Concerns\CanWinTitles;
     use Concerns\HasEmployments;
     use Concerns\HasManagers;
     use Concerns\HasMatches;
@@ -28,7 +28,6 @@ class TagTeam extends Model implements Bookable, CanBeAStableMember, Employable,
     use Concerns\HasWrestlers;
     use Concerns\OwnedByUser;
     use HasFactory;
-    use HasMorphToOne;
     use SoftDeletes;
 
     /**
