@@ -20,7 +20,7 @@ class WrestlersTableSeeder extends Seeder
         $eNum = 1;
         $now = Carbon::now();
 
-        if (null === $dateToStart) {
+        if ($dateToStart === null) {
             $dateToStart = Carbon::now()->subYears(5);
         }
 
