@@ -16,7 +16,10 @@
         <x-details-card>
             <x-card>
                 <x-card.body>
-                    <x-card.detail-link collapsibleLink="kt_tag_team_view_details" resource="tag team" :href="route('tag-teams.edit', $tagTeam)" />
+                    <x-card.detail-link
+                        collapsibleLink="kt_tag_team_view_details"
+                        resource="tag team" :href="route('tag-teams.edit', $tagTeam)"
+                    />
                     <x-separator />
                     <x-card.detail-container id="kt_tag_team_view_details">
                         <x-card.detail-row property="Name" value="{{ $tagTeam->name }}">
@@ -27,7 +30,10 @@
                             <x-card.detail-property label="Current Tag Team Partners" />
                             <x-card.detail-value>
                             @forelse ($tagTeam->currentWrestlers as $wrestler)
-                                <x-route-link :route="route('wrestlers.show', $wrestler)" label="{{ $wrestler->name }}" />
+                                <x-route-link
+                                    :route="route('wrestlers.show', $wrestler)"
+                                    label="{{ $wrestler->name }}"
+                                />
 
                                 @if ($loop->count === 1)
                                     and TBD
@@ -41,6 +47,38 @@
                             @endforelse
                             </x-card.detail-value>
                         </x-card.detail-row>
+
+                        @if ($tagTeam->currentChampionship)
+                            <x-card.detail-row>
+                                <x-card.detail-property label="Current Title Championship" />
+                                <x-card.detail-value>
+                                    <x-route-link
+                                        :route="route('titles.show', $tagTeam->currentChampionship->title)"
+                                        label="{{ $tagTeam->currentChampionship->title->name }}"
+                                    />
+                                </x-card.detail-value>
+                            </x-card.detail-row>
+                        @endif
+
+                        @if ($tagTeam->currentManagers->isNotEmpty())
+                            <x-card.detail-row>
+                                <x-card.detail-property label="Current Managers" />
+                                <x-card.detail-value>
+                                    @forelse ($tagTeam->currentManagers as $manager)
+                                        <x-route-link
+                                            :route="route('managers.show', $manager)"
+                                            label="{{ $manager->full_name }}"
+                                        />
+
+                                        @if (! $loop->last)
+                                            and
+                                        @endif
+                                    @empty
+                                        No Current Wrestlers Assigned
+                                    @endforelse
+                                </x-card.detail-value>
+                            </x-card.detail-row>
+                        @endif
 
                         @if ($tagTeam->currentWrestlers->isNotEmpty())
                             <x-card.detail-row value="{{ $tagTeam->combined_weight }} lbs.">
@@ -58,19 +96,39 @@
 
                         <x-card.detail-row>
                             <x-card.detail-property label="Start Date" />
-                            <x-card.detail-value>{{ $tagTeam->startedAt?->toDateString() ?? 'No Start Date Set' }}</x-card.detail-value>
+                            <x-card.detail-value>
+                                {{ $tagTeam->startedAt?->toDateString() ?? 'No Start Date Set' }}
+                            </x-card.detail-value>
                         </x-card.detail-row>
                     </x-card.detail-container>
 
                     @if ($tagTeam->isUnemployed())
-                        <x-notice class="mt-4" title="This tag team needs your attention!" description="This tag team does not have a start date and needs to be employed." />
+                        <x-notice
+                            class="mt-4"
+                            title="This tag team needs your attention!"
+                            description="This tag team does not have a start date and needs to be employed."
+                        />
                     @endif
                 </x-card.body>
             </x-card>
         </x-details-card>
 
         <x-details-data>
-            <livewire:tag-teams.match-list :tagTeam="$tagTeam" />
+            @if ($tagTeam->previousTitleChampionships->isNotEmpty())
+                <livewire:tag-teams.previous-title-championships-list :tagTeam="$tagTeam" />
+            @endif
+
+            @if ($tagTeam->previousMatches->isNotEmpty())
+                <livewire:tag-teams.previous-matches-list :tagTeam="$tagTeam" />
+            @endif
+
+            @if ($tagTeam->previousWrestlers->isNotEmpty())
+                <livewire:tag-teams.previous-wrestlers-list :tagTeam="$tagTeam" />
+            @endif
+
+            @if ($tagTeam->previousManagers->isNotEmpty())
+                <livewire:tag-teams.previous-managers-list :tagTeam="$tagTeam" />
+            @endif
         </x-details-data>
     </x-details-page>
 </x-layouts.app>
