@@ -16,18 +16,36 @@
         <x-details-card>
             <x-card>
                 <x-card.body>
-                    <x-card.detail-link collapsibleLink="kt_venue_view_details" resource="venue" :href="route('venues.edit', $venue)" />
+                    <x-card.detail-link
+                        collapsibleLink="kt_venue_view_details"
+                        resource="venue"
+                        :href="route('venues.edit', $venue)"
+                    />
                     <x-separator />
                     <x-card.detail-container id="kt_venue_view_details">
-                        <x-card.detail-row property="Name" value="{{ $venue->name }}" />
-                        <x-card.detail-row property="Address" value="{{ $venue->street_address }} {{ $venue->city }}, {{ $venue->state }} {{ $venue->zip }}" />
+                        <x-card.detail-row>
+                            <x-card.detail-property label="Name" />
+                            <x-card.detail-value>{{ $venue->name }}</x-card.detail-value>
+                        </x-card.detail-row>
+                        <x-card.detail-row>
+                            <x-card.detail-property label="Address" />
+                            <x-card.detail-value>
+                                {{ $venue->street_address }}
+                                @php
+                                    echo "<br />";
+                                @endphp
+                                {{ $venue->city }}, {{ $venue->state }} {{ $venue->zip }}
+                            </x-card.detail-value>
+                        </x-card.detail-row>
                     </x-card.detail-container>
                 </x-card.body>
             </x-card>
         </x-details-card>
 
         <x-details-data>
-            <livewire:venues.events-list :venue="$venue" />
+            @if ($venue->previousEvents->isNotEmpty())
+                <livewire:venues.previous-events-list :venue="$venue" />
+            @endif
         </x-details-data>
     </x-details-page>
 </x-layouts.app>
